backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
requests==2.31.0
python-jose==3.3.0
cryptography==41.0.8
qrcode==7.4.2
google-generativeai==0.3.2
